bpf: Commented out OpenBSD-specific ioctl calls
The BPF_FILDROP_CAPTURE constant and BIOCSFILDROP ioctl flag
are specific to OpenBSD. I am leaving the commented-out code
in place because FreeBSD may eventually support this ioctl
in a future release.

Perhaps an #ifndef macro would be better... The only issue with
a macro is it will consume more lines of code, potentially
increasing the likelihood of a merge conflict with upstream.