db: Do not parse lease database as root user
The lease database is parsed by dhcpd when it is running as root
(i.e., before it drops privileges). The parsing logic is not
simple, and the attacker controls some of the data that ends
up in the lease database (for example: devices' hostnames).

This commit modifies the code to open the lease database,
drop privileges, and then parse the lease database. This
change decreases the blast radius of a memory corruption
bug being exploited in the lease parsing code.